Fixed list top 3 products, moved initialization to seeders and removed bootstrap
