+++ src/main/resources/application.properties
@@ -1,4 +1,7 @@
-spring.jpa.hibernate.ddl-auto=create
+# Jost for the first application run
+# spring.jpa.hibernate.ddl-auto=create
+
+spring.jpa.hibernate.ddl-auto=update
 spring.datasource.url=jdbc:mysql://localhost:3306/bistro
 spring.datasource.username=root
 spring.datasource.password=@Ionut17122001
